Move models to db folder, use barrel export
