+++ db/index.js
@@ -1,0 +1,19 @@
+const DevProfile = require('./DevProfile');
+const Event = require('./Event');
+const ProjectReq = require('./ProjectReq');
+const Project = require('./Project');
+const User = require('./User');
+const Link = require('./Link');
+const Social = require('./Social');
+const Tag = require('./Tag');
+
+module.exports = {
+  DevProfile,
+  Event,
+  ProjectReq,
+  Project,
+  User,
+  Link,
+  Social,
+  Tag
+};
